+++ src/main/java/com/webank/webase/front/transaction/entity/ReqTransHandle.java
@@ -41,7 +41,7 @@ public class ReqTransHandle {
     private String funcName;
     private String groupId = "1";
     private List<Object> contractAbi = new ArrayList<>();
-    private List<Object> funcParam = new ArrayList<>();
+    private List<String> funcParam = new ArrayList<>();
     @Deprecated
     private boolean useAes;
     // 1.4.3